권한확인을 library 페이지로 이동
