+++ app/src/main/java/net/jspiner/epub_viewer/ui/reader/ReaderActivity.kt
@@ -1,12 +1,9 @@
 package net.jspiner.epub_viewer.ui.reader
 
-import android.Manifest
 import android.app.Activity
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
-import com.gun0912.tedpermission.PermissionListener
-import com.gun0912.tedpermission.TedPermission
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.schedulers.Schedulers
 import net.jspiner.epub_viewer.R
@@ -56,7 +53,6 @@ class ReaderActivity : BaseActivity<ActivityReaderBinding, ReaderViewModel>() {
         initViews()
 
         viewModel.setEpubFile(epubFile)
-        requestPermission()
         viewModel.getViewerType()
             .skip(1)
             .compose(bindLifecycle())
@@ -66,20 +62,6 @@ class ReaderActivity : BaseActivity<ActivityReaderBinding, ReaderViewModel>() {
     private fun initViews() {
         binding.toolboxView.touchSender = { binding.epubView.sendTouchEvent(it) }
         setNavigationBarColor(R.color.colorPrimaryDark)
-    }
-
-    private fun requestPermission() {
-        TedPermission.with(this)
-            .setPermissions(Manifest.permission.READ_EXTERNAL_STORAGE)
-            .setPermissionListener(object : PermissionListener {
-                override fun onPermissionGranted() {
-                    loadEpub()
-                }
-
-                override fun onPermissionDenied(list: MutableList<String>?) {
-                    requestPermission()
-                }
-            }).check()
     }
 
     private fun loadEpub() {
